refactor: SSE 호출 위치 변경 (Layout -> SideBar)
