+++ src/components/common/Layout/Layout.module.scss
@@ -6,4 +6,5 @@
   width: 100%;
   height: 100vh;
   @include flexSort(column, null, null, null);
+  position: relative;
 }
