+++ src/components/common/SideBar/SideBar.module.scss
@@ -7,6 +7,7 @@
 
 .header {
   height: 13.28rem;
+  @include flexSort(null, center, center, null);
 }
 
 .gridBox {
